Update profile version to 10900
Merge previous upgrade steps of this branch into it.
