--- src/brasil/gov/portal/upgrades/v10804/configure.zcml
@@ -1,30 +1,0 @@
-<configure
-    xmlns="http://namespaces.zope.org/zope"
-    xmlns:genericsetup="http://namespaces.zope.org/genericsetup">
-
-  <genericsetup:upgradeSteps
-      source="10803"
-      destination="10804"
-      profile="brasil.gov.portal:default">
-
-    <genericsetup:upgradeStep
-        title="Move styles to brasil.gov.temas package"
-        description="Move styles to brasil.gov.temas package."
-        handler=".remove_styles"
-        />
-
-    <genericsetup:upgradeStep
-        title="Show back global_sections viewlet"
-        description="Show back global_sections viewlet."
-        handler=".show_global_sections"
-        />
-
-    <genericsetup:upgradeStep
-        title="Remove collective.nitf customizations"
-        description="Remove customized view from types tool and set default view on objects using custom view."
-        handler=".remove_nitf_customizations"
-        />
-
-  </genericsetup:upgradeSteps>
-
-</configure>
