--- src/brasil/gov/portal/upgrades/v10805/configure.zcml
@@ -1,17 +1,0 @@
-<configure
-    xmlns="http://namespaces.zope.org/zope"
-    xmlns:genericsetup="http://namespaces.zope.org/genericsetup">
-
-  <genericsetup:upgradeSteps
-      source="10804"
-      destination="10805"
-      profile="brasil.gov.portal:default">
-
-    <genericsetup:upgradeStep
-        title="Remove sc.embedder from types_not_searched"
-        handler=".search_for_embedder"
-        />
-
-  </genericsetup:upgradeSteps>
-
-</configure>
