--- src/brasil/gov/portal/upgrades/v10806/configure.zcml
@@ -1,43 +1,0 @@
-<configure
-    xmlns="http://namespaces.zope.org/zope"
-    xmlns:genericsetup="http://namespaces.zope.org/genericsetup">
-
-  <genericsetup:registerProfile
-      name="default"
-      title="brasil.gov.portal (v10806)"
-      directory="profile"
-      description='Configuration for version 10806'
-      provides="Products.GenericSetup.interfaces.EXTENSION"
-      />
-
-  <genericsetup:upgradeSteps
-      source="10805"
-      destination="10806"
-      profile="brasil.gov.portal:default">
-
-    <genericsetup:upgradeDepends
-        title="Apply rolemap.xml"
-        import_steps="rolemap"
-        />
-
-    <genericsetup:upgradeDepends
-        title="Import various"
-        description="Adds new configlet and content type."
-        import_profile="brasil.gov.portal.upgrades.v10806:default"
-        />
-
-    <genericsetup:upgradeStep
-        title="Update galeria image sizes."
-        description="Update galeria de fotos image sizes."
-        handler=".update_galeria_image_sizes"
-        />
-
-    <genericsetup:upgradeStep
-        title="Install Products.PloneKeywordManager"
-        description=""
-        handler=".install_keyword_manager"
-        />
-
-  </genericsetup:upgradeSteps>
-
-</configure>
