+++ src/brasil/gov/portal/upgrades/v10900/profile/types/Infographic.xml
